[Add] dataBinding + viewModel + liveData dependencies
