+++ Kyusik/StudyClone/app/src/main/java/com/clonecodingbm/view/bottom/FavoriteFragment.kt
@@ -1,4 +1,4 @@
-package com.clonecodingbm.bottom
+package com.clonecodingbm.view.bottom
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -7,12 +7,12 @@ import android.view.View
 import android.view.ViewGroup
 import com.clonecodingbm.R
 
-class OrderListFragment : Fragment() {
+class FavoriteFragment : Fragment() {
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_order_list, container, false)
+        return inflater.inflate(R.layout.fragment_favorite, container, false)
     }
 }
