+++ Kyusik/StudyClone/app/src/main/java/com/clonecodingbm/view/bottom/MyPageFragment.kt
@@ -1,18 +1,18 @@
-package com.clonecodingbm.bottom
+package com.clonecodingbm.view.bottom
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.fragment.app.Fragment
 import com.clonecodingbm.R
 
-class FavoriteFragment : Fragment() {
+class MyPageFragment : Fragment() {
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_favorite, container, false)
+        return inflater.inflate(R.layout.fragment_my_page, container, false)
     }
 }
