+++ Kyusik/StudyClone/app/src/main/res/layout/fragment_my_page.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/teal_700"
-    tools:context=".bottom.MyPageFragment">
+    tools:context=".view.bottom.MyPageFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
