+++ Kyusik/StudyClone/app/src/main/res/layout/fragment_order_list.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".bottom.OrderListFragment">
+    tools:context=".view.bottom.OrderListFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
