+++ Kyusik/StudyClone/app/src/main/res/layout/fragment_search.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/main"
-    tools:context=".bottom.SearchFragment">
+    tools:context=".view.bottom.SearchFragment">
 
     <TextView
         android:layout_width="match_parent"
