+++ Kyusik/StudyClone/app/src/main/res/navigation/main_nav.xml
@@ -51,22 +51,30 @@
     </fragment>
     <fragment
         android:id="@+id/searchFragment"
-        android:name="com.clonecodingbm.bottom.SearchFragment"
+        android:name="com.clonecodingbm.view.bottom.SearchFragment"
         android:label="fragment_search"
         tools:layout="@layout/fragment_search" />
     <fragment
         android:id="@+id/orderListFragment"
-        android:name="com.clonecodingbm.bottom.OrderListFragment"
+        android:name="com.clonecodingbm.view.bottom.OrderListFragment"
         android:label="fragment_order_list"
         tools:layout="@layout/fragment_order_list" />
     <fragment
         android:id="@+id/myPageFragment"
-        android:name="com.clonecodingbm.bottom.MyPageFragment"
+        android:name="com.clonecodingbm.view.bottom.MyPageFragment"
         android:label="fragment_my_page"
-        tools:layout="@layout/fragment_my_page" />
+        tools:layout="@layout/fragment_my_page" >
+        <action
+            android:id="@+id/action_myPageFragment_to_loginFragment"
+            app:destination="@id/loginFragment" />
+    </fragment>
     <fragment
         android:id="@+id/favoriteFragment"
-        android:name="com.clonecodingbm.bottom.FavoriteFragment"
+        android:name="com.clonecodingbm.view.bottom.FavoriteFragment"
         android:label="fragment_favorite"
         tools:layout="@layout/fragment_favorite" />
+    <fragment
+        android:id="@+id/loginFragment"
+        android:name="com.clonecodingbm.view.mypage.LoginFragment"
+        android:label="LoginFragment" />
 </navigation>
