+++ Kyusik/StudyClone/build.gradle
@@ -1,16 +1,17 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 buildscript {
     ext.kotlin_version = "1.5.21"
-    ext.nav_version = "2.3.5"
+//    ext.nav_version = "2.3.5"
+    ext.nav_version = "2.4.0-alpha06"
+    ext.lifecycle_version = "2.4.0-alpha03"
     repositories {
         google()
         mavenCentral()
     }
     dependencies {
-        classpath "com.android.tools.build:gradle:4.2.2"
-        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
-//        classpath "androidx.navigation:navigation-safe-args-gradle-plugin:$nav_version"
-        classpath 'androidx.navigation:navigation-safe-args-gradle-plugin:2.4.0-alpha06'
+        classpath 'com.android.tools.build:gradle:7.0.2'
+        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:1.5.30"
+        classpath "androidx.navigation:navigation-safe-args-gradle-plugin:$nav_version"
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
     }
@@ -20,7 +21,6 @@ allprojects {
     repositories {
         google()
         mavenCentral()
-        jcenter() // Warning: this repository is going to shut down soon
     }
 }
